+++ docs/search/classes_1.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['counter_192',['counter',['../classenh_1_1counter.html',1,'enh']]]
+  ['counter_227',['counter',['../classenh_1_1counter.html',1,'enh']]]
 ];