+++ docs/search/classes_2.js
@@ -1,4 +1,5 @@
 var searchData=
 [
-  ['error_5fbase_193',['error_base',['../classenh_1_1error__base.html',1,'enh']]]
+  ['date_228',['date',['../classenh_1_1date.html',1,'enh']]],
+  ['datetime_229',['DateTime',['../classenh_1_1_date_time.html',1,'enh']]]
 ];