+++ docs/search/functions_9.js
@@ -1,12 +1,6 @@
 var searchData=
 [
-  ['operator_21_252',['operator!',['../classenh_1_1timer.html#a6197461f72ef9cbafa724a75bb4c4cb1',1,'enh::timer::operator!()'],['../namespaceenh.html#a812420b64dc76bb983d6b627c9e433a2',1,'enh::operator!()']]],
-  ['operator_28_29_253',['operator()',['../classenh_1_1counter.html#a968acd37a8b12b2ce16d806249ebfca6',1,'enh::counter']]],
-  ['operator_3c_254',['operator&lt;',['../classenh_1_1counter.html#af88a31ba971f3ea3567755932ff4dc06',1,'enh::counter']]],
-  ['operator_3c_3c_255',['operator&lt;&lt;',['../namespaceenh.html#a51d3bfd96daf7772870f6390df3a1d00',1,'enh']]],
-  ['operator_3c_3d_256',['operator&lt;=',['../classenh_1_1counter.html#ac6a1d3e25e9a61d996bf5c297f966150',1,'enh::counter']]],
-  ['operator_3d_257',['operator=',['../classenh_1_1counter.html#a5f92ff48cc62c0ba8784fd9c36b698ca',1,'enh::counter::operator=(const counter &amp;c) noexcept'],['../classenh_1_1counter.html#a7328668907dd222c2451794d52e46d25',1,'enh::counter::operator=(counter &amp;&amp;c) noexcept']]],
-  ['operator_3d_3d_258',['operator==',['../classenh_1_1counter.html#a77c834644c3b5cb240de0fee80353b2d',1,'enh::counter']]],
-  ['operator_3e_259',['operator&gt;',['../classenh_1_1counter.html#a54b4aa8e6462ebd5ba5984b0c1e95c3d',1,'enh::counter']]],
-  ['operator_3e_3d_260',['operator&gt;=',['../classenh_1_1counter.html#a3850e0ae998b9d3307ff7440d347394e',1,'enh::counter']]]
+  ['max_315',['max',['../namespaceenh.html#a5f461456cf52f5d925004b085dd84761',1,'enh']]],
+  ['min_316',['min',['../namespaceenh.html#af17a502dc7630b917abc0c438cbfb18c',1,'enh']]],
+  ['month_5flimit_317',['month_limit',['../namespaceenh.html#a3a89d6dc1f27d7423607ffe07f6bb3c0',1,'enh']]]
 ];