+++ docs/search/functions_a.js
@@ -1,6 +1,12 @@
 var searchData=
 [
-  ['postmessage_261',['postMessage',['../classenh_1_1queued__process.html#a894cff5b159d6e41a6069dfc6d51e93d',1,'enh::queued_process']]],
-  ['program_5felapsed_262',['program_elapsed',['../classenh_1_1timer.html#ad649090fc3529e98fba7291057413197',1,'enh::timer']]],
-  ['program_5fstart_5fpoint_263',['program_start_point',['../classenh_1_1timer.html#a190209fcccca57083a53b7e36781939f',1,'enh::timer']]]
+  ['operator_21_318',['operator!',['../classenh_1_1timer.html#a6197461f72ef9cbafa724a75bb4c4cb1',1,'enh::timer::operator!()'],['../namespaceenh.html#a812420b64dc76bb983d6b627c9e433a2',1,'enh::operator!()']]],
+  ['operator_28_29_319',['operator()',['../classenh_1_1counter.html#a968acd37a8b12b2ce16d806249ebfca6',1,'enh::counter']]],
+  ['operator_3c_320',['operator&lt;',['../classenh_1_1counter.html#af88a31ba971f3ea3567755932ff4dc06',1,'enh::counter']]],
+  ['operator_3c_3c_321',['operator&lt;&lt;',['../namespaceenh.html#a51d3bfd96daf7772870f6390df3a1d00',1,'enh']]],
+  ['operator_3c_3d_322',['operator&lt;=',['../classenh_1_1counter.html#ac6a1d3e25e9a61d996bf5c297f966150',1,'enh::counter']]],
+  ['operator_3d_323',['operator=',['../classenh_1_1counter.html#a5f92ff48cc62c0ba8784fd9c36b698ca',1,'enh::counter::operator=(const counter &amp;c) noexcept'],['../classenh_1_1counter.html#a7328668907dd222c2451794d52e46d25',1,'enh::counter::operator=(counter &amp;&amp;c) noexcept']]],
+  ['operator_3d_3d_324',['operator==',['../classenh_1_1counter.html#a77c834644c3b5cb240de0fee80353b2d',1,'enh::counter']]],
+  ['operator_3e_325',['operator&gt;',['../classenh_1_1counter.html#a54b4aa8e6462ebd5ba5984b0c1e95c3d',1,'enh::counter']]],
+  ['operator_3e_3d_326',['operator&gt;=',['../classenh_1_1counter.html#a3850e0ae998b9d3307ff7440d347394e',1,'enh::counter']]]
 ];